# doc/api.yaml
openapi: 3.0.0
paths:
  /user:
    post:
      description: This endpoint creates a new user with a login and password.
      operationId: UsersController_addUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: The user has been successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request. The input data is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - login must be longer than or equal to 3 characters
                    - password must be longer than or equal to 4 characters
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
      security: &a1
        - access-token: []
      summary: Create a new user
      tags: &a2
        - Users
    get:
      description: This endpoint retrieves a list of all users.
      operationId: UsersController_getUsers
      parameters: []
      responses:
        "200":
          description: A list of users has been successfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
      security: *a1
      summary: Get all users
      tags: *a2
  /user/{id}:
    get:
      description: This endpoint retrieves a user by their ID.
      operationId: UsersController_getUser
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the user (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "200":
          description: The user has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: User not found. The user with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: User with ID 123e4567-e89b-12d3-a456-426614174000 not found
                  error: Not Found
      security: *a1
      summary: Get a user by ID
      tags: *a2
    put:
      description: This endpoint updates the password of an existing user. The old
        password must be provided for verification.
      operationId: UsersController_updateUserPassword
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the user (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordDto"
      responses:
        "200":
          description: The user password has been successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "403":
          description: Forbidden. The user does not have permission to perform this action.
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Current password is incorrect
                  error: Forbidden
        "404":
          description: User not found. The user with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: User with ID 123e4567-e89b-12d3-a456-426614174000 not found
                  error: Not Found
      security: *a1
      summary: Update user password
      tags: *a2
    delete:
      description: This endpoint deletes a user by their ID.
      operationId: UsersController_deleteUser
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the user (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "204":
          description: The user has been successfully deleted.
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: User not found. The user with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: User with ID 123e4567-e89b-12d3-a456-426614174000 not found
                  error: Not Found
      security: *a1
      summary: Delete a user by ID
      tags: *a2
...
info:
  title: Home Library Service
  description: API for managing music albums, artists, and tracks
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: Authorization
      description: Enter JWT access token
      in: header
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          description: The login of the user
          example: john_doe
          minLength: 3
          maxLength: 20
        password:
          type: string
          description: The password of the user
          example: password
          minLength: 4
          maxLength: 100
      required:
        - login
        - password
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        login:
          type: string
          example: john_doe
        version:
          type: number
          example: 1
        createdAt:
          format: int64
          type: integer
          example: 1716927665136
        updatedAt:
          format: int64
          type: integer
          example: 1716927665136
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          description: The old password of the user
          example: password
        newPassword:
          type: string
          description: The new password of the user
          example: NewPassword
          minLength: 4
          maxLength: 100
      required:
        - oldPassword
        - newPassword
...